+++ cli/tests/test_git_init.rs
@@ -744,6 +744,79 @@ fn test_git_init_colocated_via_flag_git_dir_exists() {
 }
 
 #[test]
+fn test_git_init_colocated_import_branches() {
+    let test_env = TestEnvironment::default();
+
+    let origin_root = test_env.env_root().join("origin");
+    init_git_repo(&origin_root, true);
+
+    let workspace_root = test_env.env_root().join("repo_track");
+    git2::Repository::clone(origin_root.to_str().unwrap(), &workspace_root).unwrap();
+
+    let (stdout, stderr) = test_env.jj_cmd_ok(
+        test_env.env_root(),
+        &[
+            "git",
+            "init",
+            "--colocate",
+            "repo_track",
+            "--config",
+            "git.init-track-local-bookmarks=true",
+        ],
+    );
+
+    insta::assert_snapshot!(stdout, @"");
+    insta::assert_snapshot!(stderr, @r#"
+    Done importing changes from the underlying Git repo.
+    Setting the revset alias "trunk()" to "my-bookmark@origin"
+    Tracking the following remote bookmarks:
+      my-bookmark@origin
+    Initialized repo in "repo_track"
+    "#);
+
+    // Check that the bookmark is checked out
+    let stdout = test_env.jj_cmd_success(&workspace_root, &["bookmark", "list", "--all"]);
+    insta::assert_snapshot!(stdout, @r#"
+    my-bookmark: mwrttmos 8d698d4a My commit message
+      @git: mwrttmos 8d698d4a My commit message
+      @origin: mwrttmos 8d698d4a My commit message
+    "#);
+
+    let workspace_root = test_env.env_root().join("repo_no_track");
+    git2::Repository::clone(origin_root.to_str().unwrap(), &workspace_root).unwrap();
+
+    let (stdout, stderr) = test_env.jj_cmd_ok(
+        test_env.env_root(),
+        &[
+            "git",
+            "init",
+            "--colocate",
+            "repo_no_track",
+            "--config",
+            "git.init-track-local-bookmarks=false",
+        ],
+    );
+
+    insta::assert_snapshot!(stdout, @"");
+    insta::assert_snapshot!(stderr, @r#"
+    Done importing changes from the underlying Git repo.
+    Setting the revset alias "trunk()" to "my-bookmark@origin"
+    Hint: The following remote bookmarks aren't associated with the existing local bookmarks:
+      my-bookmark@origin
+    Hint: Run `jj bookmark track my-bookmark@origin` to keep local bookmarks updated on future pulls.
+    Initialized repo in "repo_no_track"
+    "#);
+
+    // Check that the bookmark is *not* checked out
+    let stdout = test_env.jj_cmd_success(&workspace_root, &["bookmark", "list", "--all"]);
+    insta::assert_snapshot!(stdout, @r#"
+    my-bookmark: mwrttmos 8d698d4a My commit message
+      @git: mwrttmos 8d698d4a My commit message
+    my-bookmark@origin: mwrttmos 8d698d4a My commit message
+    "#);
+}
+
+#[test]
 fn test_git_init_colocated_via_flag_git_dir_not_exists() {
     let test_env = TestEnvironment::default();
     let workspace_root = test_env.env_root().join("repo");